+++ src/main/java/com/example/bank/security/config/ApplicationConfig.java
@@ -37,7 +37,7 @@ public class ApplicationConfig {
     }
 
     @Bean
-    public AuditorAware<Integer> auditorAware() {
+    public AuditorAware<Long> auditorAware() {
         return new ApplicationAuditAware();
     }
